Text patch (NY00123)
Two main engine-side changes:
- id_vh* and id_vl* have new methods, letting you "xor" graphics onto the virtual screen. In the SDL 1.2 case the screen has XRGB format, so this may be a bit of a problem if the palette is changed. But for now it doesn't happen, so we have the current code (and of course SDL 2.0).
- Three functions in id_vh.c, including VH_MeasurePropString, have been modified to pick the correct font chunk (offset of 3 added).
- Only US_GetPrintFont, US_GetPrintColour and the respective setters are used. (Would be much simpler to access the variables directly, although less clean in a way, I agree.)

+++ src/id_us.h
@@ -69,16 +69,16 @@ int US_GetWindowW();
 int US_GetWindowH();
 int US_GetPrintX();
 int US_GetPrintY();
-int US_GetPrintFont();
-int US_GetPrintColour();
+int16_t US_GetPrintFont();
+int8_t US_GetPrintColour();
 void US_SetWindowX(int parm);
 void US_SetWindowY(int parm);
 void US_SetWindowW(int parm);
 void US_SetWindowH(int parm);
 void US_SetPrintX(int parm);
 void US_SetPrintY(int parm);
-void US_SetPrintFont(int parm);
-void US_SetPrintColour(int parm);
+void US_SetPrintFont(int16_t parm);
+void US_SetPrintColour(int8_t parm);
 
 // Common
 extern bool us_noWait; // Debug mode enabled.
